offset2: split data journal and offset out of log
just moving some code around for clearer organisation
