+++ offset2/log.go
@@ -1,9 +1,7 @@
 package offset2 // import "go.cryptoscope.co/margaret/offset2"
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
 	"io"
 	"os"
 	"path/filepath"
@@ -197,168 +195,6 @@ func (log *offsetLog) checkConsistency() error {
 			return errors.Errorf("offset mismatch: offset file says %d, data file has %d", expOfst, ofst)
 		}
 	}
-}
-
-type data struct {
-	*os.File
-
-	buf [8]byte
-}
-
-func (d *data) frameReader(ofst int64) (io.Reader, error) {
-	var sz int64
-	err := binary.Read(io.NewSectionReader(d, ofst, 8), binary.BigEndian, &sz)
-	if err != nil {
-		return nil, errors.Wrap(err, "error reading payload length")
-	}
-
-	return io.NewSectionReader(d, ofst+8, sz), nil
-}
-
-func (d *data) readFrame(data []byte, ofst int64) (int, error) {
-	sr := io.NewSectionReader(d, ofst, 8)
-
-	var sz int64
-	err := binary.Read(sr, binary.BigEndian, &sz)
-	if err != nil {
-		return 0, errors.Wrap(err, "error reading payload length")
-	}
-
-	return d.ReadAt(data, ofst+8)
-}
-
-func (d *data) getFrameSize(ofst int64) (int64, error) {
-	_, err := d.ReadAt(d.buf[:], ofst)
-	if err != nil {
-		return 0, errors.Wrap(err, "error reading payload length")
-	}
-
-	buf := bytes.NewBuffer(d.buf[:])
-
-	var sz int64
-	err = binary.Read(buf, binary.BigEndian, &sz)
-	return sz, errors.Wrap(err, "error parsing payload length")
-}
-
-func (d *data) getFrame(ofst int64) ([]byte, error) {
-	sz, err := d.getFrameSize(ofst)
-	if err != nil {
-		return nil, errors.Wrap(err, "error getting frame size")
-	}
-
-	data := make([]byte, sz)
-	_, err = d.readFrame(data, ofst)
-	return data, errors.Wrap(err, "error reading frame")
-}
-
-func (d *data) append(data []byte) (int64, error) {
-	ofst, err := d.Seek(0, io.SeekEnd)
-	if err != nil {
-		return 0, errors.Wrap(err, "failed to seek to end of file")
-	}
-
-	err = binary.Write(d, binary.BigEndian, int64(len(data)))
-	if err != nil {
-		return 0, errors.Wrap(err, "writing length prefix failed")
-	}
-
-	_, err = d.Write(data)
-	return ofst, errors.Wrap(err, "error writing data")
-}
-
-type offset struct {
-	*os.File
-}
-
-func (o *offset) readOffset(seq margaret.Seq) (int64, error) {
-	_, err := o.Seek(seq.Seq()*8, io.SeekStart)
-	if err != nil {
-		return 0, errors.Wrap(err, "seek failed")
-	}
-
-	var ofst int64
-	err = binary.Read(o, binary.BigEndian, &ofst)
-	return ofst, errors.Wrap(err, "error reading offset")
-}
-
-func (o *offset) readLastOffset() (int64, margaret.Seq, error) {
-	stat, err := o.Stat()
-	if err != nil {
-		return 0, margaret.SeqEmpty, errors.Wrap(err, "stat failed")
-	}
-
-	sz := stat.Size()
-	if sz == 0 {
-		return 0, margaret.SeqEmpty, nil
-	}
-
-	// this should be off-by-one-error-free:
-	// sz is 8 when there is one entry, and the first entry has seq 0
-	seqOfst := margaret.BaseSeq(sz/8 - 1)
-
-	var ofstData int64
-	err = binary.Read(io.NewSectionReader(o, sz-8, 8), binary.BigEndian, &ofstData)
-	if err != nil {
-		return 0, margaret.SeqEmpty, errors.Wrap(err, "error reading entry")
-	}
-
-	return ofstData, seqOfst, nil
-}
-
-func (o *offset) append(ofst int64) (margaret.Seq, error) {
-	ofstOfst, err := o.Seek(0, io.SeekEnd)
-	seq := margaret.BaseSeq(ofstOfst / 8)
-	if err != nil {
-		return seq, errors.Wrap(err, "could not seek to end of file")
-	}
-
-	err = binary.Write(o, binary.BigEndian, ofst)
-	return seq, errors.Wrap(err, "error writing offset")
-}
-
-type journal struct {
-	*os.File
-}
-
-func (j *journal) readSeq() (margaret.Seq, error) {
-	stat, err := j.Stat()
-	if err != nil {
-		return margaret.SeqEmpty, errors.Wrap(err, "stat failed")
-	}
-
-	switch sz := stat.Size(); sz {
-	case 0:
-		return margaret.SeqEmpty, nil
-	case 8:
-		// continue after switch
-	default:
-		return margaret.SeqEmpty, errors.Errorf("expected file size of 8B, got %dB", sz)
-	}
-
-	_, err = j.Seek(0, io.SeekStart)
-	if err != nil {
-		return margaret.SeqEmpty, errors.Wrap(err, "could not seek to start of file")
-	}
-
-	var seq margaret.BaseSeq
-	err = binary.Read(j, binary.BigEndian, &seq)
-	return seq, errors.Wrap(err, "error reading seq")
-}
-
-func (j *journal) bump() (margaret.Seq, error) {
-	seq, err := j.readSeq()
-	if err != nil {
-		return margaret.SeqEmpty, errors.Wrap(err, "error reading old journal value")
-	}
-
-	_, err = j.Seek(0, io.SeekStart)
-	if err != nil {
-		return margaret.SeqEmpty, errors.Wrap(err, "could not seek to start of file")
-	}
-
-	seq = margaret.BaseSeq(seq.Seq() + 1)
-	err = binary.Write(j, binary.BigEndian, seq)
-	return seq, errors.Wrap(err, "error writing seq")
 }
 
 func (log *offsetLog) Seq() luigi.Observable {
